// client/src/data/Region.ts
enum Region {
	SubsaharanAfrica = "Sub-Saharan Africa",
	MiddleEast = "Middle East",
	EasternEurope = "Eastern Europe",
	LatinAmerica = "Latin America",
	India = "India",
	SouthAsia = "South Asia",
	China = "China",
	Japan = "Japan",
	CentralJava = "Central Java",
	Bali = "Bali",
}

export default Region;